==> 1. Features.Rmd <==

```{r}
library(skimr)
library(tidyverse)
library(tidymodels)
library(janitor)
library(dplyr)
```


```{r}
train <- read_csv('C:/Users/ASUS/Desktop/MATHia/training_set_with_formatted_time.csv')
```

```{r}
summary <- read_csv('C:/Users/ASUS/Desktop/MATHia/workspace_summary_train.csv')
summary <- summary %>% distinct()
```


# Get features from the "training_set_with_formatted_time" file

```{r}
df <- train %>%
  group_by(Anon.Student.Id) %>%
  summarise(
    
    # 1. # Number of problem completed 
    num_problem = n_distinct(`Problem.Name`), 
    
    # 2. JIT Feedback Per Problem
    avg_jit_per_problem = sum(grepl("JIT", Outcome)) / num_problem,
    
    # 3. Total Attempts Per Problem
    avg_attempt_per_problem = ifelse(num_problem > 0, sum(Action == "Attempt") / num_problem, 0),
    
    # 4. Correct Attempts Per Problem
    avg_correct_attempt_per_problem = ifelse(
      num_problem > 0,
      sum(Outcome %in% c("OK", "OK_AMBIGUOUS")) / num_problem,
      0
    ),
    
    # 5. Correct Attempt Ratio
    correct_attempt_ratio = sum(Outcome %in% c("OK", "OK_AMBIGUOUS")) /sum(Action == "Attempt"),
    
    # 6. avg_hint_count
    avg_hint_count = sum(Outcome %in% c("INITIAL_HINT", "HINT_LEVEL_CHANGE")) /num_problem,
    
    # 7. avg_error_count
    avg_error_count = sum(Outcome == "ERROR") /num_problem
  
  ) %>%
  ungroup()

```

# 8. Graduation rate
```{r}
graduation_rate <- train %>%
  distinct(Anon.Student.Id, Problem.Name, .keep_all = TRUE) %>% 
  group_by(Anon.Student.Id) %>%
  summarise(
    num_problem = n(),  # Number of problems
    graduation_rate = sum(CF..Workspace.Progress.Status. == "GRADUATED") / num_problem
  ) %>%
  dplyr::select(-num_problem)
```


```{r}
df_merged <- df %>%
  left_join(graduation_rate, by = "Anon.Student.Id")
```


```{r}
#write.csv(df, "C:/Users/ASUS/Desktop/MATHia/features.csv", row.names = FALSE)
```


# Get features from the summary file

```{r}
df_summary <- summary %>% 
  group_by(Anon.Student.Id) %>%
  summarise(
    # 9. avg_time_per_workspace
    
    avg_time_per_workspace = mean(workspace_total_time_seconds),
    
    # 10.  avg_step_problems_completed
    avg_step_problems_completed = mean(step_by_step_problems_completed),
    
    # 11. avg_problem_completion_ratio
    avg_problem_completion_ratio = mean(problems_completed / max_problem_count),
    
    # 12. Skills
    avg_skill_mastery_ratio = mean(
      if_else(skills_encountered > 0,
              skills_mastered / skills_encountered,
              1)
    ),
    
    # 13. aplse
    aplse_score_ratio = mean(aplse_earned / aplse_possible)
  )
```

# merge together
```{r}
df_merged <- df_merged %>%
  left_join(df_summary, by = "Anon.Student.Id")
```


# combine with hints ------------

```{r}
hint_counts <- train %>%
  filter(Action == "Hint Request") %>%
  group_by(Anon.Student.Id) %>%
  summarise(
    num_problem = n_distinct(`Problem.Name`), 
    
    # 14. hint 1
    Level1_Hints_Per_Problem = sum(Help.Level == 1) / num_problem,
    
    # 15. hint 2
    Level2_Hints_Per_Problem = sum(Help.Level == 2) / num_problem,
    
    # 16. hint 3
    Level3_Hints_Per_Problem = sum(Help.Level == 3) / num_problem
    ) %>%
  dplyr::select(-num_problem)
```

# merge again 
```{r}
df_merged <- df_merged %>%
  left_join(hint_counts, by = "Anon.Student.Id")
```



#===========================#
#     Output File           #
#===========================#

```{r}
write.csv(df_merged, "C:/Users/ASUS/Desktop/MATHia/features.csv", row.names = FALSE)
```
